[HM] | feature-create-room-feature -add feature to create-room

+++ src/main/java/io/github/hotelmanagement/model/room/RoomDTO.java
@@ -10,8 +10,7 @@ public record RoomDTO(
         Long id,
         int bedAmount,
         int maxPeopleInside,
-        int pricePerNight,
-
+        double pricePerNight,
         List<ReservationDTO> reservationDTOS
 )
 {
